requirements.txt
asgiref==3.3.1
cffi==1.14.4
Django==2.2
django-cors-headers==3.6.0
djangorestframework==3.12.2
Pillow==8.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
